backend/requirements.txt
# Filename: requirements.txt
# Location: backend/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.1

# LangChain and Agent Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pydantic==2.5.3
